Debugging and refactoring: StrikeScore
debugging showed the error 'BowlingGame' object has no attribute 'StrikeScore'. after looking through the 'BowlingGame' script I noticed the 'StrikeScore' function was mispelled as 'stickeScore'. I also changed the initial 'StrikeScore' to start with a lower case s so it follows the same layout as the rest of the functions