Debugging and refactoring: testGutterGame
first I tried to diagnose the testGutterBall function as it would be the simplest to fix. In the score() function, I set up some print() functions to check what parts are being called, and it turns out the gutter ball was being counted as a strike, so I realized the check "if frameIndex in range(10)" is illogical as the 'frameIndex' variable is used to count the for loop, so it will always be within the range. instead I followed the similar coding of 'self.isSpare()' found below but with the unused 'isStrike()' function to check whether the roll is a strike. after testing this the testGutterBall test passed with no errors.