Debugging and refactoring: testAllOnes
next I tried to diagnose the error on the 'testAllOnes()' test. to debug this I used the print function on several areas of the score, but then I realised that despite having a for loop, the loop was only running once. then I realised the 'return restult' was placed inside the loop, rendering it pointless. I since moved it out of the for loop, down to the end of the function. after this the 'testAllOnes()' test passed without errors