Refactoring
After the debugging was performed and the tests passed successfully, I performed some refactoring to improve the code while maintaining its core function. Due to how simple the code was, I believed the only part in need of improvement was the naming convention. I changed each variable and function from its previous method of using Capitals to separate words, and instead used underscores (_) to separate words in variable and function names as this is the python naming convention for variable and function names.

I additionally made a requirement so the code will call an error if an invalid input is given to the roll function. This seems counterproductive to the prior efforts to remove errors, but I believe this will be useful for potential future maintenance to fully accept or deny a requested action.